+++ flod/lib/src/rules/numbers/base_number_rule.dart
@@ -1,7 +1,7 @@
 import 'package:flod/src/rules/base_rule.dart';
 
 abstract class BaseNumberRule<T extends num> extends BaseRule<T> {
-  const BaseNumberRule({required super.code});
+  const BaseNumberRule({required super.code, super.message});
 
   Map<String, dynamic> get params => const {};
 }