+++ flod/lib/src/rules/numbers/custom_number_rule.dart
@@ -7,6 +7,7 @@ class CustomNumberRule<T extends num> extends BaseNumberRule<T> {
   const CustomNumberRule(
     this.predicate, {
     required super.code,
+    super.message,
     this.metaParams,
   });