+++ flod/lib/src/rules/regexp/regex_rule.dart
@@ -3,7 +3,7 @@ import 'package:flod/src/rules/string/base_string_rule.dart';
 class RegexRule extends BaseStringRule {
   final RegExp pattern;
 
-  const RegexRule(this.pattern, {required super.code});
+  const RegexRule(this.pattern, {required super.code, super.message});
 
   @override
   bool check(String value) => pattern.hasMatch(value);