+++ flod/lib/src/rules/string/base_string_rule.dart
@@ -1,7 +1,7 @@
 import 'package:flod/src/rules/base_rule.dart';
 
 abstract class BaseStringRule extends BaseRule<String> {
-  const BaseStringRule({required super.code});
+  const BaseStringRule({required super.code, super.message});
 
   Map<String, dynamic> get params => const {};
 }