+++ flod/lib/src/rules/string/fixed_length_rule.dart
@@ -3,7 +3,7 @@ import 'package:flod/src/rules/string/base_string_rule.dart';
 class FixedLengthRule extends BaseStringRule {
   final int length;
 
-  FixedLengthRule(this.length, {required super.code});
+  FixedLengthRule(this.length, {required super.code, super.message});
 
   @override
   bool check(String value) => value.length == length;