+++ flod/lib/src/validators/nullable_and_optional_validator/optional_validator.dart
@@ -60,6 +60,7 @@ class OptionalValidator<T> extends Validator<T?> with Transformable<T?> {
                 path: e.path,
                 code: e.code,
                 params: e.params,
+                message: e.message,
                 value: null,
                 isSecret: true,
               ),